Migrate pools + wallet routes
